# Conflicts:
#	backend/Origam.DA.Service/AbstractSqlDataService.cs
#	backend/Origam.DA.Service/Origam.DA.Service.csproj
#	backend/Origam.Rule/RuleEngine.cs
#	backend/Origam.Rule/Xslt/AbstractXsltEngine.cs
#	backend/Origam.Server/Configuration/StartUpConfiguration.cs
#	backend/Origam.Server/Startup.cs
#	backend/Origam.Workflow/Engine.cs
#	backend/Origam.sln

+++ frontend-html/src/model/actions-ui/DataView/TableView/onFieldKeyDown.ts
@@ -33,12 +33,18 @@ import {
   isAddRecordShortcut,
   isDeleteRecordShortcut,
   isDuplicateRecordShortcut,
-  isFilterRecordShortcut
+  isFilterRecordShortcut, isRefreshShortcut
 } from "utils/keyShortcuts";
 import { onDeleteRowClick } from "model/actions-ui/DataView/onDeleteRowClick";
 import { onCreateRowClick } from "model/actions-ui/DataView/onCreateRowClick";
 import { onCopyRowClick } from "model/actions-ui/DataView/onCopyRowClick";
 import { onFilterButtonClick } from "model/actions-ui/DataView/onFilterButtonClick";
+import { onEscapePressed } from "model/actions-ui/DataView/onEscapePressed";
+import { getFormScreenLifecycle } from "model/selectors/FormScreen/getFormScreenLifecycle";
+import { getScreenActionButtonsState } from "model/actions-ui/ScreenToolbar/saveButtonVisible";
+import { getIsAddButtonVisible } from "model/selectors/DataView/getIsAddButtonVisible";
+import { getIsDelButtonVisible } from "model/selectors/DataView/getIsDelButtonVisible";
+import { getIsCopyButtonVisible } from "model/selectors/DataView/getIsCopyButtonVisible";
 
 export function onFieldKeyDown(ctx: any) {
 
@@ -51,6 +57,7 @@ export function onFieldKeyDown(ctx: any) {
     try {
       const dataView = getDataView(ctx);
       const tablePanelView = getTablePanelView(ctx);
+      const gridFocusManager = getGridFocusManager(tablePanelView);
       tablePanelView.handleEditorKeyDown(event);
       switch (event.key) {
         case "Tab": {
@@ -62,13 +69,13 @@ export function onFieldKeyDown(ctx: any) {
               return;
             }
 
-            yield dataView.lifecycle.runRecordChangedReaction(function*() {
-              if (event.shiftKey) {
-                yield selectPrevColumn(ctx)(true);
-              } else {
-                yield selectNextColumn(ctx)(true);
-              }
-            });
+            if (event.shiftKey) {
+              yield*selectPrevColumn(ctx)(true);
+            } else {
+              yield*selectNextColumn(ctx)(true);
+            }
+            yield*dataView.lifecycle.runRecordChangedReaction();
+
             event.preventDefault();
             tablePanelView.dontHandleNextScroll();
             tablePanelView.scrollToCurrentCell();
@@ -76,16 +83,16 @@ export function onFieldKeyDown(ctx: any) {
             tablePanelView.triggerOnFocusTable();
           } else {
             if (event.shiftKey) {
-              selectPrevColumn(ctx)(true);
+              yield*selectPrevColumn(ctx)(true);
             } else {
-              selectNextColumn(ctx)(true);
+              yield*selectNextColumn(ctx)(true);
             }
             event.preventDefault();
 
             tablePanelView.dontHandleNextScroll();
             tablePanelView.scrollToCurrentCell();
             yield*flushCurrentRowData(ctx)();
-            getGridFocusManager(ctx).focusEditor();
+            gridFocusManager.focusEditor();
           }
           break;
         }
@@ -100,40 +107,61 @@ export function onFieldKeyDown(ctx: any) {
             return;
           }
 
-          yield dataView.lifecycle.runRecordChangedReaction(function*() {
-            if (event.shiftKey) {
-              yield*selectPrevRow(ctx)();
-            } else {
-              yield*selectNextRow(ctx)();
-            }
-          });
+          if (event.shiftKey) {
+            yield*selectPrevRow(ctx)();
+          } else {
+            yield*selectNextRow(ctx)();
+          }
+          yield*dataView.lifecycle.runRecordChangedReaction();
 
-          tablePanelView.setEditing(true);
           tablePanelView.scrollToCurrentCell();
+          tablePanelView.setEditing(true);
           break;
         }
         case "F2": {
           tablePanelView.setEditing(false);
+          gridFocusManager.activeEditor = undefined;
           tablePanelView.triggerOnFocusTable();
           break;
         }
         case "Escape": {
-          tablePanelView.setEditing(false);
-          tablePanelView.clearCurrentCellEditData();
-          tablePanelView.triggerOnFocusTable();
+          if(!event.closedADropdown){
+            yield onEscapePressed(dataView, event);
+            tablePanelView.setEditing(false);
+            gridFocusManager.activeEditor = undefined;
+            tablePanelView.clearCurrentCellEditData();
+            tablePanelView.triggerOnFocusTable();
+          }
           break;
         }
         default: {
           if (isSaveShortcut(event)) {
+            yield*flushCurrentRowData(ctx)();
+            if (getScreenActionButtonsState(ctx)?.isSaveButtonVisible) {
+              const formScreenLifecycle = getFormScreenLifecycle(ctx);
+              yield*formScreenLifecycle.onSaveSession();
+            }
+          }
+          else if (isRefreshShortcut(event) && getScreenActionButtonsState(ctx)?.isRefreshButtonVisible) {
             tablePanelView.setEditing(false);
             yield*flushCurrentRowData(ctx)();
-          } else if (isAddRecordShortcut(event)) {
+            const formScreenLifecycle = getFormScreenLifecycle(ctx);
+            yield*formScreenLifecycle.onRequestScreenReload();
+          }
+          else if (isAddRecordShortcut(event) && getIsAddButtonVisible(dataView)) {
+            tablePanelView.setEditing(false);
+            yield*flushCurrentRowData(ctx)();
             yield onCreateRowClick(dataView)(event);
-          } else if (isDeleteRecordShortcut(event)) {
+          }
+          else if (isDeleteRecordShortcut(event) && getIsDelButtonVisible(dataView)) {
             yield onDeleteRowClick(dataView)(event);
-          } else if (isDuplicateRecordShortcut(event)) {
+          }
+          else if (isDuplicateRecordShortcut(event) && getIsCopyButtonVisible(dataView)) {
+            tablePanelView.setEditing(false);
+            yield*flushCurrentRowData(ctx)();
             yield onCopyRowClick(dataView)(event);
-          } else if (isFilterRecordShortcut(event)) {
+          }
+          else if (isFilterRecordShortcut(event)) {
             yield onFilterButtonClick(dataView)(event);
           }
         }
